Moved BitsInput and BitsOutput from comm package to io package, as they are not exclusive to comm streams.
